refactor: move `OrderEvent` into `portfolio/mod.rs`
