+++ src/event.rs
@@ -1,6 +1,6 @@
 use crate::data::MarketEvent;
 use crate::execution::FillEvent;
-use crate::portfolio::order::OrderEvent;
+use crate::portfolio::OrderEvent;
 use crate::portfolio::position::{EquityPoint, Position, PositionExit, PositionUpdate};
 use crate::portfolio::repository::{AvailableCash, TotalEquity};
 use crate::strategy::signal::{SignalEvent, SignalForceExit};
